+++ src/main/java/com/example/surveyapp/domain/order/infrastructure/OrderJpaRepository.java
@@ -1,4 +1,4 @@
-package com.example.surveyapp.domain.order.infrastructor;
+package com.example.surveyapp.domain.order.infrastructure;
 
 import com.example.surveyapp.domain.order.domain.model.Order;
 import org.springframework.data.domain.Page;
